html/semantics/embedded-content-0/media-elements/loading-the-media-resource/020.html

<!doctype html>
<title>invoking resource selection by inserting into a document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var v = document.createElement('video');
  document.body.appendChild(v);
  assert_equals(v.networkState, 3, 'networkState after document.body.appendChild(v)');
  window.onload = t.step_func(function() {
    assert_equals(v.networkState, 0, 'networkState in window.onload');
    t.done();
  });
});
</script>
